refactor: 폴더 이름 변경 및 아이콘 저장 (FE-53)
